+++ NEWS.md
@@ -1,9 +1,9 @@
 gridpattern v0.1
 ================
 
-* ``grid.pattern()`` draws a specified "pattern" to a graphic device.
+* ``grid.pattern()`` draws a specified "pattern" to the active graphics device.
 * ``patternGrob()`` returns a "pattern" (grid) grob.
-* Initial support for the following patterns from ``ggpattern``:
+* Initial support for the following patterns from [ggpattern](https://github.com/coolbutuseless/ggpattern):
 
   1. [circle](https://coolbutuseless.github.io/package/ggpattern/articles/pattern-circle.html) (#5)
   2. [crosshatch](https://coolbutuseless.github.io/package/ggpattern/articles/pattern-crosshatch.html) (#4)
@@ -11,3 +11,4 @@ gridpattern v0.1
   4. [magick](https://coolbutuseless.github.io/package/ggpattern/articles/pattern-magick.html) (#6)
   5. [stripe](https://coolbutuseless.github.io/package/ggpattern/articles/pattern-stripe.html) (#3)
   6. [Custom geometry-based patterns](https://coolbutuseless.github.io/package/ggpattern/articles/developing-patterns-2.html) (#1)
+  7. [Custom array-based patterns](https://coolbutuseless.github.io/package/ggpattern/articles/developing-patterns-3.html) (#2)